+++ src/Controller/CategoryController.php
@@ -25,6 +25,10 @@ class CategoryController extends AbstractController
     #[Route('/new', name: 'category_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+		if (!$this->getUser()) {
+			return $this->render('forbidden.html.twig');
+		}
+		
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
@@ -53,6 +57,10 @@ class CategoryController extends AbstractController
     #[Route('/{id}/edit', name: 'category_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Category $category, EntityManagerInterface $entityManager): Response
     {
+		if (!$this->getUser() || $this->getUser()->getId() != 1) {
+			return $this->render('forbidden.html.twig');
+		}
+		
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
@@ -71,6 +79,10 @@ class CategoryController extends AbstractController
     #[Route('/{id}', name: 'category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category, EntityManagerInterface $entityManager): Response
     {
+		if (!$this->getUser() || $this->getUser()->getId() != 1) {
+			return $this->render('forbidden.html.twig');
+		}
+		
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
             $entityManager->remove($category);
             $entityManager->flush();
